git-svn-id: http://47.100.187.235:13932/svn/metadatamanager/metadatamanager@384 3051ac0c-2819-405f-8e91-413e4011bda0

+++ src/main/java/cn/sowell/datacenter/model/dictionary/service/impl/BasicItemServiceImpl.java
@@ -151,7 +151,14 @@ public class BasicItemServiceImpl implements BasicItemService {
             
         } else if (String.valueOf(ValueType.RECORD.getIndex()).equals(basicItem.getOneLevelItem().getDataType())) {
         	BasicItem lableObj = basicItemDao.getLableObj(basicItem.getCode());
-        	basicItemDao.delete(lableObj);
+        	if (lableObj != null) {
+        		basicItemDao.delete(lableObj);
+        	}
+        	//删除实体编辑时间属性
+        	BasicItem bt = basicItemDao.get(BasicItem.class, AttributeParter.getLeafEditTimeName(basicItem.getCode()));
+        	if (bt != null) {
+        		basicItemDao.delete(bt);
+        	}
         }
 		
 		basicItemDao.delete(basicItem);
